% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_arima.R
\name{fit_agg_arima}
\alias{fit_agg_arima}
\title{fit_agg_arima}
\usage{
fit_agg_arima(
  formula,
  x,
  k,
  orders = rep(NA, length(x)),
  seasonal = rep(NA, length(x)),
  m = rep(NA, length(x)),
  mean = TRUE,
  fixed = NULL,
  ...
)
}
\arguments{
\item{formula}{formula object for modeling the data. if no regressor, use intercept only.}

\item{x}{list of data.frames, containing variables according to formula}

\item{k}{vector of aggregation levels}

\item{orders}{list of ARIMA orders per level of aggregation}

\item{seasonal}{list of seasonal ARIMA orders per level of aggregation}

\item{m}{vector of frequencies. required if seasonal order is given}

\item{mean}{include mean (T/F)}

\item{fixed}{list of vectors to fix parameters in the models}

\item{...}{additional parameters for model function}
}
\value{
object of class agg_arima containing
\itemize{
\item models
\item formula
\item k
\item orders
\item fixed
\item xtest_idx
\item data x
}
}
\description{
fit_agg_arima
}
\details{
Each entry of x must be a data.frame with columns index, corresponding to the observed timestep. This should be aligned across all levels.
The formula must align with the columns available in the data.frame, i.e. if response is x with regressor a, then formula = x~a.

Fit ARIMA models to multiple aggregation levels of a time series
}
