% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_arima.R
\name{fitted2}
\alias{fitted2}
\title{Obtain fitted values for agg. ARIMA models in a more efficient way}
\usage{
fitted2(object, steps, h = 1)
}
\arguments{
\item{object}{result of fit_agg_arima}

\item{steps}{number of new data steps steps for each level of the hierarchy. Useful to incorporate on-line data.}

\item{h}{horizon used to internally compute fitted values (only h=1 allowed so far)}
}
\value{
list of data.frames with variables index and x
}
\description{
Obtain fitted values for agg. ARIMA models in a more efficient way
}
