% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_demand_daily}
\alias{food_demand_daily}
\title{Daily Food Demand Dataset}
\format{
A tibble of size 40,464 x 5 containing
\itemize{
\item fridge_id
\item date
\item day
\item delivered
\item sold
}
}
\usage{
food_demand_daily
}
\description{
Daily time series of smart fridges, containing delivered and sold items.
}
\keyword{datasets}
