% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_and_reconcile.R
\name{forecast_and_reconcile}
\alias{forecast_and_reconcile}
\title{forecast_and_reconcile}
\usage{
forecast_and_reconcile(
  formula,
  x,
  k,
  model,
  var_type = c("var", "struc", "series", "bu"),
  cov_type = c("ols", "scaling", "full", "cov_shrink", "autocov_scaling",
    "markov_scaling", "inv_glasso", "corr_shrink", "spectral", "bu"),
  h = 1,
  model_train_prop = 0.8,
  recon_train_prop = 0,
  err_measure = function(y, yhat = 0) (mean((y - yhat)^2, na.rm = TRUE)),
  test_recon = c("non_cum", "ex_ante", "ex_post"),
  h_test = NULL,
  ...
)
}
\arguments{
\item{formula}{formula object for modeling the data. if no regressor, use intercept only.}

\item{x}{list of data.frames, containing variables according to formula}

\item{k}{vector of aggregation levels}

\item{model}{list with fit function and possible params for the fit function}

\item{var_type}{type of variance to be used on each aggregation level}

\item{cov_type}{type of covariance to be used between various aggregation levels}

\item{h}{forecast horizon based on top level of the hierarchy}

\item{model_train_prop}{proportion of x to be used for model fitting}

\item{recon_train_prop}{proportion of x to be used for only reconciliation purposes (must be 0, not used)}

\item{err_measure}{error measure to be used in possible cross-validation as well as for calculating train/test errors}

\item{test_recon}{how should reconciliation be handled on the test data? non-cumulatively, ex-ante or ex-post?}

\item{h_test}{fixed test set size. required if model_train_prop is NULL}

\item{...}{additional parameters for reconcile_forecasts()}
}
\value{
object of class fit_recon with
\itemize{
\item fits output of fit_agg_arima
\item y training data
\item ytest test data
\item yhat training base forecasts
\item yhat_test test base forecasts
\item k aggregation parameter
\item result named list of recon object for each reconciliation type
}
}
\description{
Fit ARIMA models to each level of temporal hierarchy and reconcile those base forecasts using selected reconciliation methods
}
\details{
in contrast to reconcile_forecasts, multiple choices for var_type, cov_type are allowed.
}
\seealso{
\code{\link[=fit_agg_arima]{fit_agg_arima()}}, \code{\link[=reconcile_forecasts]{reconcile_forecasts()}}
}
