% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nem_generation_by_source}
\alias{nem_generation_by_source}
\title{Australian Electricity Generation}
\format{
A tibble of size 8,418 × 3 containing
\itemize{
\item date
\item Source
\item Generation
}
}
\source{
\url{https://github.com/PuwasalaG/Probabilistic-Forecast-Reconciliation}, \url{http://opennem.org.au}
}
\usage{
nem_generation_by_source
}
\description{
Daily time series of electricity generation in Australia per source.
}
\references{
Panagiotelis, A., Gamakumara, P., Athanasopoulos, G., and Hyndman, R. J. (2023). Probabilistic forecast reconciliation: Properties, evaluation and score optimisation. European Journal of Operational Research, 306(2):693–706.
}
\keyword{datasets}
