% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.R
\name{plot.agg_ts}
\alias{plot.agg_ts}
\title{Plot temporally aggregated ts}
\usage{
\method{plot}{agg_ts}(object, type = c("ts", "acf", "pacf"), ...)
}
\arguments{
\item{object}{of class agg_ts}

\item{type}{of plot. ts to plot trajectories, acf/pacf to plot (partial-) autocorrelation functions}

\item{...}{not used}
}
\value{
ggplot2 object
}
\description{
Plot temporally aggregated ts
}
