% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_and_reconcile.R
\name{plot.fit_recon}
\alias{plot.fit_recon}
\title{Plot the time series, base and reconciled forecasts}
\usage{
\method{plot}{fit_recon}(obj, facet_fmla = ~var_type + cov_type, filter = NULL, est_mapping = NULL)
}
\arguments{
\item{obj}{object of class fit_recon}

\item{facet_fmla}{control the facetting of the plot, e.g. should the reconciled forecasts across all method be displayed in one plot?}

\item{filter}{list with names var_type, cov_type to filter recon. methods to plot}

\item{est_mapping}{named vector to give recon. methods proper names}
}
\value{
ggplot2 object
}
\description{
Plot the time series, base and reconciled forecasts
}
