% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_arima.R
\name{predict.agg_arima}
\alias{predict.agg_arima}
\title{Predict/Forecast agg. ARIMA models}
\usage{
\method{predict}{agg_arima}(object, h = 1, partial = TRUE, newxreg = NULL)
}
\arguments{
\item{object}{of class agg_arima}

\item{h}{forecast horizon based on top level of the aggregation hierarchy}

\item{partial}{allow partially observed hierarchies (TRUE some lower level forecasts may be actual observations, FALSE replace partial observations with fitted values)}

\item{newxreg}{list of regressors if initially used to fit the models}
}
\value{
list of data.frames with variables pred and obs, indicating if the forecasts is an actual forecast or not
}
\description{
Predict/Forecast agg. ARIMA models
}
