% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_arima.R
\name{print.agg_arima}
\alias{print.agg_arima}
\title{Print aggregated ARIMA models}
\usage{
\method{print}{agg_arima}(x, ...)
}
\arguments{
\item{x}{result of fit_agg_arima function}

\item{...}{additional parameters to print of individual ARIMA models}
}
\description{
Print aggregated ARIMA models
}
