% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prison_population}
\alias{prison_population}
\title{Australian Prison Population}
\format{
A tibble of size 3,072 x 6 containing
\itemize{
\item Date
\item State
\item Gender
\item Legal
\item Indigenous
\item Count
}
}
\source{
\url{https://otexts.com/fpp3/extrafiles/prison_population.csv}
}
\usage{
prison_population
}
\description{
Quarterly time series of prison population in Australia per various categories.
}
\references{
Hyndman, R. J., & Athanasopoulos, G. (2018). Forecasting: principles and practice. OTexts.
}
\keyword{datasets}
