% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile_forecasts}
\alias{reconcile_forecasts}
\title{reconcile_forecasts}
\usage{
reconcile_forecasts(
  y,
  yhat,
  k,
  var_type = c("var", "struc", "series", "bu"),
  cov_type = c("ols", "scaling", "full", "cov_shrink", "autocov_scaling",
    "markov_scaling", "inv_glasso", "corr_shrink", "spectral", "bu"),
  yhat_test = NULL,
  test_recon_cum = FALSE,
  ytest = NULL,
  S = NULL,
  G = NULL,
  SG = NULL,
  full_cov = NULL,
  sntz = FALSE,
  ...
)
}
\arguments{
\item{y}{data matrix representing the hierarchy of time series}

\item{yhat}{data matrix of base forecasts with same dimension as y}

\item{k}{vector of aggregation levels}

\item{var_type}{type of variance to be used on each aggregation level}

\item{cov_type}{type of covariance to be used between various aggregation levels}

\item{yhat_test}{optional base test forecasts with same number of columns as y and yhat}

\item{test_recon_cum}{should the reonciliation method be re-estimated cumulatively on the test data (T/F)?}

\item{ytest}{data matrix of test data in order to be able to use test_recon_cum=TRUE}

\item{S}{summing matrix. if not given, it is computed according to k}

\item{G}{mapping matrix. if not given, it is estimated by selected reconciliation method}

\item{SG}{transformation matrix}

\item{full_cov}{if not given, it is estimated using data y and yhat}

\item{...}{additional parameters for selected reconciliation methods}
}
\value{
list with
\itemize{
\item yhat_tilde matrix of reconciled forecasts with same dimension as yhat
\item cov_est
\item S
\item G
\item yhat_test_tilde matrix of reconciled test forecasts with same dimension as yhat_test (if given)
}
}
\description{
Reconcile base forecasts using selected reconciliation method
}
\details{
cov_shrink and corr_shrink types need additional lambda parameter between 0 and 1.
inv_glasso requires rho>0 parameter.
spectral scaling requires lambda and neig parameters.

if multiple are given, then cv is performed to find optimal one.
}
\seealso{
\code{\link[=reconcile_forecasts.cv]{reconcile_forecasts.cv()}}
}
