% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile_partly.R
\name{reconcile_forecasts_partly}
\alias{reconcile_forecasts_partly}
\title{reconcile_forecasts_partly}
\usage{
reconcile_forecasts_partly(
  fits,
  bottom_steps = 1L,
  var_type = c("var", "struc", "series", "bu"),
  cov_type = c("ols", "scaling", "full", "cov_shrink", "autocov_scaling",
    "markov_scaling", "inv_glasso", "corr_shrink", "spectral", "bu"),
  err_measure = function(y, yhat = 0) mean((y - yhat)^2, na.rm = TRUE),
  ytest = NULL,
  full_cov = NULL,
  fix = c("model", "order"),
  sntz = FALSE,
  x = NULL,
  xreg = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{fits}{a fitted hierarchical model with fitted, fitted2, update, predict, update_and_predict functions and fields k (and data x)}

\item{bottom_steps}{number of bottom level steps, must be greater or equal to 1}

\item{var_type}{type of variance to be used on each aggregation level}

\item{cov_type}{type of covariance to be used between various aggregation levels}

\item{err_measure}{error measure function of form f(y,yhat)}

\item{ytest}{list of test data to compute test errors}

\item{full_cov}{if not given, it is estimated using fits}

\item{fix}{for model updating fix either "model" or "order"}

\item{sntz}{use set-negative-to-zero heuristic (T/F)}

\item{x}{optional list containining the data used to model the fits}

\item{xreg}{optional regressors, if not given, should be contained in x (either by fits or given)}

\item{verbose}{detailed output (T/F)}

\item{...}{additional parameters for selected reconciliation methods}
}
\value{
tibble with columns
\itemize{
\item var_type
\item cov_type
\item level
\item \strong{train_hat_err} base forecast training errors
\item \strong{train_tilde_err} reconciled training errors for no additional data
\item \strong{train_hat_now_err} updated base forecast training errors
\item \strong{train_tilde_now_err} updated reconciled training errors
\item \strong{test_hat_err} base forecast test errors
\item \strong{test_tilde_err} reconciled test errors for no additional data
\item \strong{test_hat_now_err} updated base forecast test errors
\item \strong{test_tilde_now_err} updated reconciled test errors
\item \strong{train_tilde_now_base_rel_err} relative errors train_tilde_now_err/train_hat_err - 1
\item \strong{train_tilde_base_rel_err} relative errors train_tilde_err/train_hat_err - 1
\item \strong{train_tilde_now_rel_err} relative errors train_tilde_now_err/train_tilde_err - 1
\item \strong{test_tilde_now_base_rel_err} relative errors test_tilde_now_err/test_hat_err - 1
\item \strong{test_tilde_rel_err} relative errors test_tilde_err/test_hat_err - 1
\item \strong{test_tilde_now_rel_err} relative errors test_tilde_now_err/test_tilde_err - 1
}
if verbose, then also included list-columns are: hat, test_hat, tilde, test_tilde, hat_now, test_hat_now, tilde_now, test_tilde_now, obs_dat, SG, SG_now
}
\description{
Reconcile base forecasts using selected reconciliation method in a partly manner
}
\details{
This function can be used to perform temporal hierarchical forecast updating by using partial information of the hierarchy.

cov_shrink and corr_shrink types need additional lambda parameter between 0 and 1.
inv_glasso requires rho>0 parameter.
spectral scaling requires lambda and neig parameters.

if multiple are given, then cv is performed to find optimal one.
}
\seealso{
\code{\link[=reconcile_forecasts]{reconcile_forecasts()}}
}
