% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spower}
\alias{spower}
\title{Solar Power Generation}
\format{
A data.frame of size 8,198,502 x 4 containing
\itemize{
\item State
\item CapacityMW
\item LocalTime
\item Power
}
}
\usage{
spower
}
\description{
Simulated photovoltaic generation data per state and capacity in MW, on a 5-min basis for the year 2006.
}
\keyword{datasets}
