% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_and_reconcile.R
\name{summary.fit_recon}
\alias{summary.fit_recon}
\title{Print summary of forecast_and_reconcile output. It prints the summary of the fits as well as a selection of errors (which is also returned invisibly)}
\usage{
\method{summary}{fit_recon}(object, ...)
}
\arguments{
\item{object}{object of class fit_recon}

\item{...}{not used}
}
\description{
Print summary of forecast_and_reconcile output. It prints the summary of the fits as well as a selection of errors (which is also returned invisibly)
}
