% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_arima.R
\name{update.agg_arima}
\alias{update.agg_arima}
\title{Update aggregated ARIMA models with new data}
\usage{
\method{update}{agg_arima}(object, newdata, fix = c("model", "order"))
}
\arguments{
\item{object}{result of fit_agg_arima}

\item{newdata}{list of newdata, with same format used for initially fitting the models. It is also possible to only update parts of the aggregation hierarchy (in an on-line setting)}

\item{fix}{what to fix in the models (models or only the orders)}
}
\value{
updated agg_arima object
}
\description{
Update aggregated ARIMA models with new data
}
