% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_arima.R
\name{update_and_predict}
\alias{update_and_predict}
\title{update_and_predict}
\usage{
update_and_predict(
  object,
  newdata_mat,
  h = 1,
  cumulative = TRUE,
  fix = c("model", "order"),
  newxreg = NULL,
  ...
)
}
\arguments{
\item{object}{of class agg_arima}

\item{newdata_mat}{matrix of new data with number of columns = size of hierarchy}

\item{h}{forecast horizon based on top level of the hierarchy}

\item{cumulative}{perform cumulative one-step ahead forecasts or multi-step forecast (T/F)}

\item{fix}{what to fix in the models (models or only the orders)}

\item{newxreg}{regressors if initially used in fitting the models}

\item{...}{not used}
}
\value{
list of predict data.frames
}
\description{
Update aggregated ARIMA models with new data and forecast after
}
\seealso{
\code{\link[=update.agg_arima]{update.agg_arima()}} \code{\link[=predict.agg_arima]{predict.agg_arima()}}
}
