% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VNdata}
\alias{VNdata}
\title{Australian Nightly Visitors}
\format{
A time series of size 228 x 525 containing with each column representing a region in Australia.
}
\source{
\url{https://robjhyndman.com/publications/mint/}, \url{https://github.com/daniGiro/ctprob}
}
\usage{
VNdata
}
\description{
Monthly time series of nightly visitors in various regions in Australia.
}
\references{
Wickramasuriya, S. L., Athanasopoulos, G., and Hyndman, R. J. (2019). Optimal forecast reconciliation for hierarchical and grouped time series through trace minimization. Journal of the American Statistical Association, 114(526):804–819.

Girolimetto, D., Athanasopoulos, G., Di Fonzo, T., and Hyndman, R. J. (2023). Cross-temporal probabilistic forecast reconciliation: Methodological and practical issues. International Journal of Forecasting.
}
\keyword{datasets}
